+++ src/parsers/kufar/categories/other.ts
@@ -1,12 +1,12 @@
-import { IAd, ICollection } from '../../tasks/parseKufar';
+import { type IAd } from 'config/types';
 
 const TITLE_AD_SELECTOR = 'h3[class^="styles_title__"]';
 const PRICE_AD_SELECTOR = 'p[class^=styles_price__] span';
 const IMAGE_URL_AD_SELECTOR = 'img[class^="styles_image__"]';
 const COMPANY_AD_SELECTOR = 'div[class^="styles_badge__"]';
 
-export function parserOthers(items: NodeListOf<Element>): ICollection<IAd> {
-  const newAds: ICollection<IAd> = {};
+export function parserOthers(items: NodeListOf<Element>): IAd[] {
+  const newAds: IAd[] = [];
 
   items.forEach((node) => {
     const urlItem = node.querySelector('a')?.getAttribute('href') ?? '';
@@ -18,7 +18,7 @@ export function parserOthers(items: NodeListOf<Element>): ICollection<IAd> {
       node.querySelector(TITLE_AD_SELECTOR)?.textContent?.trim() ?? '';
     const imgUrlAd =
       node.querySelector(IMAGE_URL_AD_SELECTOR)?.getAttribute('src') ??
-      'assets/no-photo.webp';
+      'https://i.ibb.co/NLkvZYG/no-photo.webp';
     const priceAd =
       node
         .querySelector(PRICE_AD_SELECTOR)
@@ -27,14 +27,13 @@ export function parserOthers(items: NodeListOf<Element>): ICollection<IAd> {
       !node.querySelector(COMPANY_AD_SELECTOR)?.textContent;
 
     if (isNotCompanyAd) {
-      newAds[itemIdAd] = {
+      newAds.push({
         img_url: imgUrlAd,
         id: itemIdAd,
         title: titleAd,
         price: priceAd,
         url: urlAd,
-        createAd: new Date().toLocaleDateString('ru-RU'),
-      };
+      });
     }
   });
   return newAds;
